Fix linux error

+++ ACC_K.cpp
@@ -297,7 +297,7 @@ double ACC_K::intervalQuery(unsigned int item, int i, int j)
     if (i == 0)
       sum = winQuery(itemIdx, j);
     else
-      sum = abs(winQuery(itemIdx, j) - winQuery(itemIdx, i));
+      sum = abs(double(winQuery(itemIdx, j) - winQuery(itemIdx, i)));
 	}
 
 	return threshold * (sum + 2);
